Remove a feasible-file-length-p restriction
This is an important choice - all this while, we had guard verification
making sure we weren't going to put together some nonsense based on the
file length and the file contents being wildly out of sync. This
scenario is now possible in an execution, since guard verification won't
catch it. However, we now don't have to put in an l6-fsck clause in the
guard of l6-rdchs, which would have been ridiculous. Also, we can model
what happens when we try to read from an ill-formed file system.