Fix a co-simulation test
I have to admit, there's some voodoo involved in this fix. The
underlying insight is: the actual number of clusters in the data region
differs from the number of entries allowed in the FAT, because the FAT
has to occupy an integral number of sectors, and will most likely be
longer than it strictly needs to be. I figured this out during my
swim (well, OK, on the way to the swimming pool) and I'm pretty stoked.